moved user fields to a seaprate component
