+++ app/components/admin/lessons/LessonFields.js
@@ -3,22 +3,21 @@ import connectToStores from 'alt-utils/lib/connectToStores';
 import {
     Row, Col, FormGroup,
     FormControl, InputGroup,
-    HelpBlock, ControlLabel, Glyphicon,
-    Well, Button
+    HelpBlock, ControlLabel, Glyphicon
 } from 'react-bootstrap';
-import { Typeahead } from 'react-bootstrap-typeahead';
+
+
+import LessonAttendees from './Attendees';
 
 import AuthStore from '../../../stores/AuthStore';
 import LessonStore from '../../../stores/LessonStore';
 import LessonActions from '../../../actions/LessonActions';
 import GroupActions from '../../../actions/GroupActions';
 import LocationActions from '../../../actions/LocationActions';
-import UserActions from '../../../actions/UserActions';
 
 import { maxInputLength } from '../../../utils/config';
 import {
-    initDateTimePicker, handleDateChange,
-    constructUserInfo, constructUserOptions
+    initDateTimePicker, handleDateChange
 } from '../../../utils/utils';
 
 /**
@@ -31,17 +30,6 @@ class LessonFields extends React.Component {
 
     static getPropsFromStores() {
         return LessonStore.getState();
-    }
-
-    /**
-     * Retrieved group members (i.e. user data)
-     *
-     * @public
-     */
-    componentWillMount() {
-        const { token } = AuthStore.getState();
-
-        UserActions.getList(token, this._userListReceived.bind(this));
     }
 
     /**
@@ -79,35 +67,6 @@ class LessonFields extends React.Component {
     }
 
     /**
-     * Mark that a user will be attending the lesson.
-     *
-     * Clears the typeahead when the user had been selected.
-     * Uses setTimeout of 1 ms to put in on the next thing on the event loop.
-     *
-     * @public
-     * @param {Object} updatable Updatable lesson
-     * @param {Object} selected selected option
-     */
-    addAttendee(updatable, selected) {
-        if (!selected || !selected[0] ||!selected[0].id) {
-            return
-        } else {
-            setTimeout(() => this.typeahead.getInstance().clear(), 1);
-        }
-
-        const userId = selected[0].id;
-
-        if (updatable.attendees.indexOf(userId) >= 0) {
-            return toastr.error('Lietotājs jau nodarbībai pievienots!');
-        }
-
-        updatable.attendees.push(userId);
-
-        LessonActions.setUpdatable(updatable);
-        LessonActions.attendeeAdded(userId);
-    }
-
-    /**
      * Render a group select option.
      *
      * @public
@@ -126,85 +85,6 @@ class LessonFields extends React.Component {
                 {groupInfo}
             </option>
         );
-    }
-
-    /**
-     * Renders a button with the user's info and adds a click event handler for removing the user from the attendance list.
-     *
-     * @public
-     * @param {Object} updatable Updatable group
-     * @param {string} attendee Lesson attendee (user) ID
-     * @param {number} index Lesson attendee index
-     * @returns {string} HTML markup
-     */
-    renderAttendee(updatable, attendee, index) {
-        const { userList } = this.props;
-        const user = userList.find(user => user._id === attendee);
-        let attendeeInfo = '';
-
-        if (user) {
-            const { given_name, family_name, email } = user;
-            attendeeInfo = constructUserInfo(email, given_name, family_name);
-        }
-
-        return (
-            <Button
-                key={`LessonAttendee${index}`}
-                bsSize="small"
-                onClick={this.removeAttendee.bind(this, updatable, attendee)}>
-                {attendeeInfo} <Glyphicon glyph="remove" />
-            </Button>
-        );
-    }
-
-    /**
-     * Renders a user select option.
-     *
-     * @public
-     * @param {Object} user User to select
-     * @param {number} index User index
-     * @returns {string} HTML markup
-     */
-    renderUser(user, index) {
-        const { _id, given_name, family_name, email } = user;
-        const userInfo = constructUserInfo(email, given_name, family_name);
-
-        return (
-            <option
-                key={`UserOption${index}`}
-                value={_id}>
-                {userInfo}
-            </option>
-        );
-    }
-
-    /**
-     * Removes an attendee from a lesson.
-     *
-     * @public
-     * @param {Object} updatable Updatable object
-     * @param {string} attendee Lesson attendee ID
-     */
-    removeAttendee(updatable, attendee) {
-        const attendeeIndex = updatable.attendees.indexOf(attendee);
-
-        updatable.attendees.splice(attendeeIndex, 1);
-
-        LessonActions.setUpdatable(updatable);
-        LessonActions.attendeeRemoved(attendee);
-    }
-
-    /**
-     * User list received event handler.
-     *
-     * Sets the user list store property.
-     *
-     * @private
-     * @param userList
-     * @private
-     */
-    _userListReceived(userList) {
-        LessonActions.userListReceived(userList);
     }
 
     /**
@@ -229,9 +109,8 @@ class LessonFields extends React.Component {
      * @returns {string} HTML markup
      */
     render() {
-        const { updatable, groups, locations, userList } = this.props;
-        const { group, location, comment, attendees } = updatable;
-        const userOptions = constructUserOptions(userList);
+        const { updatable, groups, locations } = this.props;
+        const { group, location, comment } = updatable;
 
         return (
             <div>
@@ -333,28 +212,7 @@ class LessonFields extends React.Component {
                         </FormGroup>
                     </Col>
                 </Row>
-                <Row>
-                    <Col xs={12}>
-                        <FormGroup>
-                            <ControlLabel>Pievienot nodarbības dalībnieku</ControlLabel>
-                            <Typeahead
-                                multiple={false}
-                                options={userOptions}
-                                onChange={this.addAttendee.bind(this, updatable)}
-                                placeholder="Pievienot dalībnieku"
-                                ref={ ref => this.typeahead = ref }
-                            />
-                        </FormGroup>
-                    </Col>
-                </Row>
-                <Row>
-                    <Col xs={12}>
-                        <ControlLabel>Nodarbības dalībnieki</ControlLabel>
-                        <Well bsSize="small">
-                            {attendees.map((attendee, index) => this.renderAttendee(updatable, attendee, index))}
-                        </Well>
-                    </Col>
-                </Row>
+                <LessonAttendees />
             </div>
         );
     }
